+++ app/controllers/events_controller.rb
@@ -1,4 +1,5 @@
 class EventsController < ApplicationController
+  before_action :find_user_character_ids, only: %i[index]
   before_action :find_event, only: %i[show]
 
   def index; end
@@ -10,6 +11,10 @@ class EventsController < ApplicationController
   def new; end
 
   private
+
+  def find_user_character_ids
+    @user_character_ids = Current.user.characters.pluck(:id)
+  end
 
   def find_event
     @event = Event.find_by(slug: params[:id])