+++ app/serializers/event_serializer.rb
@@ -1,5 +1,5 @@
 class EventSerializer < ActiveModel::Serializer
-  attributes :id, :name, :date, :time, :slug, :fraction_id
+  attributes :id, :name, :date, :time, :slug, :fraction_id, :description, :dungeon_id, :owner_id
 
   def date
     object.start_time.strftime('%-d.%-m.%Y')